Fix compiation bug on qv4l2
From: Mauro Carvalho Chehab <mchehab@infradead.org>

sys/time.h is required to include videodev2.h at userspace.
	thanks to audetto@tiscalli.it for pointing it.

Also fixed a warning at a printf.

+++ utils/qv4l2/general-tab.cpp
@@ -295,7 +295,7 @@ void GeneralTab::updateStandard()
 			tvStandard->setCurrentItem(vs.index);
 			what.sprintf("TV Standard (0x%llX)\n"
 				"Frame period: %f (%d/%d)\n"
-				"Frame lines: %d\n", std,
+				"Frame lines: %d\n", (long long int)std,
 				(double)vs.frameperiod.numerator / vs.frameperiod.denominator,
 				vs.frameperiod.numerator, vs.frameperiod.denominator,
 				vs.framelines);
